// healthin-app/src/app/shared/interfaces/exam-realized.ts
export interface ExamRealized {
    id: string;
    examId: string;
    examName: string;
    patientId: string;
    askData: string;
    askDoctorId: string;
    askDoctorName: string;
    liberationDate: string;
    liberationEmployeeId: string;
    liberationEmployeeName: string;
    doData: string;
    doDoctorId: string;
    doDoctorName: string;
    contribution: number;
}